swarm-os/requirements.txt
streamlit>=1.28.0
aiohttp>=3.9.0
networkx>=3.1
pyvis>=0.3.2
openai>=1.3.0
python-dotenv>=1.0.0
